requirements.txt
langchain
langchain_core
langchain_groq
pandas
python-dotenv
langchain_astradb
langchain_google_genai
fastapi
uvicorn
jinja2
python-multipart
-e.